added an image for the MyOwnPart into the simple example
This is done via the already in use and injected ResorucePool of the PartEditor parent class.
To extend this an IResourceProviderService is installed which creates and returns the image
for a given name. The BasicResourceProvider is able to create Images and Colors via defining
service properties with key is the value you use inside your code and the value is either
the path to an image or a color definition (e.g.: rgb(126,123,167)).